Replaced SW25 with Rotary Encoder. Takes place of SW25 in matrix
Removed the footprint for switch 25 and replaced it with an Alps EC11E encoder. Encoder outputs have been connected to A3(PF4) and A2(PF5). The switch part of the encoder connects to the old spot of SW25 in the matrix. I've used a modified version of the footprint Keeb.io made for this particular encoder and added it to the lily58 footprint library. Also plotted gerbers, stored in the "_gerbers" folder.